% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{acceptance.rate}
\alias{acceptance.rate}
\title{Calculate the MCMC acceptance rate}
\usage{
acceptance.rate(obj)
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}
}
\description{
Calculate the MCMC acceptance rate
}

