% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{as.mcmc.fiducialSpline.knownNumberKnots}
\alias{as.mcmc.fiducialSpline.knownNumberKnots}
\title{Convert a fiducialSpline object to a mcmc.list}
\usage{
as.mcmc.fiducialSpline.knownNumberKnots(obj)
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}
}
\description{
Convert a fiducialSpline object to a mcmc.list
}

