% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{calc.start.point}
\alias{calc.start.point}
\title{Calculate a reasonable starting point for the fiducial MCMC free-knot b-spline}
\usage{
calc.start.point(x, y, degree, knots)
}
\arguments{
\item{x}{A vector of x-values}

\item{y}{A vector of y-values}

\item{degree}{The degree of b-spline}

\item{knots}{A vector of initial knot points}
}
\description{
Calculate a reasonable starting point for the fiducial MCMC free-knot b-spline
}

