% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{confint.fiducialSpline}
\alias{confint.fiducialSpline}
\title{Calculate the fiducial confidence intervals (credible intervals???)}
\usage{
\method{confint}{fiducialSpline}(obj, level = 0.95)
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}

\item{level}{The level of the interval}
}
\description{
Calculate the fiducial confidence intervals (credible intervals???)
}

