% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{E.spline}
\alias{E.spline}
\title{Predict the value of the function at the given x-values}
\usage{
E.spline(x, spline)
}
\arguments{
\item{x}{Vector of x-values}

\item{spline}{An object produced by fiducialSpline}
}
\description{
Predict the value of the function at the given x-values
}

