% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{plot.fiducialSpline}
\alias{plot.fiducialSpline}
\title{Plot the fiducial spline}
\usage{
\method{plot}{fiducialSpline}(obj, grid.length = 401, xlab = "")
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}

\item{grid.length}{The number of points along the x-axis to predict at.}

\item{xlab}{The label of the x-axis}
}
\description{
Plot the fiducial spline
}

