% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{thin}
\alias{thin}
\title{Thin the MCMC chains}
\usage{
thin(obj, by = 10)
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}

\item{by}{The thinning interval}
}
\description{
Thin the MCMC chains
}

