% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FiducialSpline.R
\name{trim}
\alias{trim}
\title{Trim the beginning of the MCMC chains}
\usage{
trim(obj, amount)
}
\arguments{
\item{obj}{A fiducialSpline object created using fiducialSpline}

\item{amount}{Number of initial steps to remove}
}
\description{
Trim the beginning of the MCMC chains
}

